requirements.txt
# Core dependencies
pydantic>=2.0.0
pyyaml>=6.0
aiohttp>=3.8.0
rich>=13.0.0

# Provider dependencies
openai>=1.0.0
tiktoken>=0.5.0
anthropic>=0.18.0

# AWS support (optional)
boto3>=1.28.0

# Azure support (optional)
azure-ai-contentsafety>=1.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.4.0
isort>=5.12.0
